requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
aiomysql>=0.1.0
beautifulsoup4>=4.10.0
table2ascii>=1.0.0
# Add prettier or black if you use them for formatting
